+++ tests/evaluation_test.v
@@ -13,7 +13,7 @@ fn test_evaluation_single_digit_number() {
 	expected := 5.0
 
 	actual := ast.evaluate(p.parse()!)!
-	assert actual == expected, assertion_failed_msg(actual, f32(expected))
+	assert actual == expected, assertion_failed_msg(actual, expected)
 }
 
 fn test_evaluation_addition() {
@@ -25,7 +25,7 @@ fn test_evaluation_addition() {
 	expected := 69.0
 
 	actual := ast.evaluate(p.parse()!)!
-	assert actual == expected, assertion_failed_msg(actual, f32(expected))
+	assert actual == expected, assertion_failed_msg(actual, expected)
 }
 
 fn test_evaluation_addition_with_float_values() {
@@ -37,7 +37,7 @@ fn test_evaluation_addition_with_float_values() {
 	expected := 10.0
 
 	actual := ast.evaluate(p.parse()!)!
-	assert actual == expected, assertion_failed_msg(actual, f32(expected))
+	assert actual == expected, assertion_failed_msg(actual, expected)
 }
 
 fn test_evaluation_subtraction() {
@@ -49,7 +49,7 @@ fn test_evaluation_subtraction() {
 	expected := 20.0
 
 	actual := ast.evaluate(p.parse()!)!
-	assert actual == expected, assertion_failed_msg(actual, f32(expected))
+	assert actual == expected, assertion_failed_msg(actual, expected)
 }
 
 fn test_evaluation_subtraction_with_negative_result() {
@@ -61,7 +61,7 @@ fn test_evaluation_subtraction_with_negative_result() {
 	expected := -3.0
 
 	actual := ast.evaluate(p.parse()!)!
-	assert actual == expected, assertion_failed_msg(actual, f32(expected))
+	assert actual == expected, assertion_failed_msg(actual, expected)
 }
 
 fn test_evaluation_multiplication() {
@@ -73,7 +73,7 @@ fn test_evaluation_multiplication() {
 	expected := 14.0
 
 	actual := ast.evaluate(p.parse()!)!
-	assert actual == expected, assertion_failed_msg(actual, f32(expected))
+	assert actual == expected, assertion_failed_msg(actual, expected)
 }
 
 fn test_evaluation_division() {
@@ -85,7 +85,7 @@ fn test_evaluation_division() {
 	expected := 2.0
 
 	actual := ast.evaluate(p.parse()!)!
-	assert actual == expected, assertion_failed_msg(actual, f32(expected))
+	assert actual == expected, assertion_failed_msg(actual, expected)
 }
 
 fn test_evaluation_division_non_integer_result() {
@@ -97,7 +97,7 @@ fn test_evaluation_division_non_integer_result() {
 	expected := 0.5
 
 	actual := ast.evaluate(p.parse()!)!
-	assert actual == expected, assertion_failed_msg(actual, f32(expected))
+	assert actual == expected, assertion_failed_msg(actual, expected)
 }
 
 fn test_evaluation_correct_order_of_operations() {
@@ -109,7 +109,7 @@ fn test_evaluation_correct_order_of_operations() {
 	expected := 5.0
 
 	actual := ast.evaluate(p.parse()!)!
-	assert actual == expected, assertion_failed_msg(actual, f32(expected))
+	assert actual == expected, assertion_failed_msg(actual, expected)
 }
 
 fn test_evaluation_correct_order_of_operations_with_parenthesis() {
@@ -121,7 +121,7 @@ fn test_evaluation_correct_order_of_operations_with_parenthesis() {
 	expected := -2.5
 
 	actual := ast.evaluate(p.parse()!)!
-	assert actual == expected, assertion_failed_msg(actual, f32(expected))
+	assert actual == expected, assertion_failed_msg(actual, expected)
 }
 
 fn test_evaluation_with_parenthesis() {
@@ -133,7 +133,7 @@ fn test_evaluation_with_parenthesis() {
 	expected := -7.0
 
 	actual := ast.evaluate(p.parse()!)!
-	assert actual == expected, assertion_failed_msg(actual, f32(expected))
+	assert actual == expected, assertion_failed_msg(actual, expected)
 }
 
 fn test_evaluation_with_multiple_parenthesis() {
@@ -145,7 +145,7 @@ fn test_evaluation_with_multiple_parenthesis() {
 	expected := -1.0
 
 	actual := ast.evaluate(p.parse()!)!
-	assert actual == expected, assertion_failed_msg(actual, f32(expected))
+	assert actual == expected, assertion_failed_msg(actual, expected)
 }
 
 fn test_evaluation_with_nested_parenthesis() {
@@ -157,7 +157,7 @@ fn test_evaluation_with_nested_parenthesis() {
 	expected := 7.0
 
 	actual := ast.evaluate(p.parse()!)!
-	assert actual == expected, assertion_failed_msg(actual, f32(expected))
+	assert actual == expected, assertion_failed_msg(actual, expected)
 }
 
 fn test_evaluation_with_multiple_nested_parenthesis() {
@@ -169,5 +169,5 @@ fn test_evaluation_with_multiple_nested_parenthesis() {
 	expected := 2.0
 
 	actual := ast.evaluate(p.parse()!)!
-	assert actual == expected, assertion_failed_msg(actual, f32(expected))
+	assert actual == expected, assertion_failed_msg(actual, expected)
 }
